adaptado DAO de projetos
